Split Media into a public API Media<'a> and private API MLine
